refactor: move insight cash flow types
