+++ frontend/src/insights/types/cashFlow.ts
@@ -1,0 +1,9 @@
+export type CashFlowGranularity = 'day' | 'week' | 'month'
+
+export type CashFlowBarBucket = {
+  label: string
+  rangeLabel: string
+  inflow: number
+  outflow: number
+  net: number
+}
